# src/api/rest/docs/swagger.yaml
basePath: /tumblebug
definitions:
...
  model.CustomImageStatus:
    enum:
    - Available
    - Unavailable
    type: string
    x-enum-varnames:
    - MyImageAvailable
    - MyImageUnavailable
...
  model.KeyValue:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
...
  model.SimpleMsg:
    properties:
      message:
        example: Any message
        type: string
    type: object
...
  model.TbCustomImageInfo:
    properties:
      associatedObjectList:
        items:
          type: string
        type: array
      connectionName:
        example: aws-ap-southeast-1
        type: string
      creationDate:
        example: "2022-10-18T08:12:48Z"
        type: string
      cspCustomImageId:
        example: ami-06eb41e14121c550a
        type: string
      cspCustomImageName:
        example: aws-ap-southeast-1-snapshot
        type: string
      description:
        type: string
      guestOS:
        description: Windows7, Ubuntu etc.
        type: string
      id:
        example: aws-ap-southeast-1-snapshot
        type: string
      isAutoGenerated:
        type: boolean
      keyValueList:
        items:
          $ref: '#/definitions/model.KeyValue'
        type: array
      name:
        example: aws-ap-southeast-1-snapshot
        type: string
      namespace:
        description: required to save in RDB
        example: default
        type: string
      sourceVmId:
        example: aws-ap-southeast-1-1
        type: string
      status:
        allOf:
        - $ref: '#/definitions/model.CustomImageStatus'
        example: Available
      systemLabel:
        description: SystemLabel is for describing the Resource in a keyword (any
          string can be used) for special System purpose
        example: Managed by CB-Tumblebug
        type: string
      uuid:
        description: uuid is universally unique identifier for the resource
        type: string
    type: object
...
  model.TbVmSnapshotReq:
    properties:
      name:
        example: aws-ap-southeast-1-snapshot
        type: string
    type: object
...
info:
  contact:
    name: API Support
    url: https://github.com/cloud-barista/cb-tumblebug/issues/new/choose
  description: CB-Tumblebug is an open source system for managing multi-cloud infrastructure
    consisting of resources from multiple cloud service providers. (Cloud-Barista)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/cloud-barista/cb-tumblebug/blob/main/README.md
  title: CB-Tumblebug REST API
  version: latest
paths:
...
  /ns/{nsId}/control/mci/{mciId}:
    get:
      consumes:
      - application/json
      description: Control the lifecycle of MCI (refine, suspend, resume, reboot,
        terminate)
      operationId: GetControlMci
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - default: mci01
        description: MCI ID
        in: path
        name: mciId
        required: true
        type: string
      - description: Action to MCI
        enum:
        - suspend
        - resume
        - reboot
        - terminate
        - refine
        - continue
        - withdraw
        in: query
        name: action
        required: true
        type: string
      - description: Force control to skip checking controllable status
        enum:
        - "false"
        - "true"
        in: query
        name: force
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Control the lifecycle of MCI (refine, suspend, resume, reboot, terminate)
      tags:
      - '[MC-Infra] MCI Provisioning and Management'
  /ns/{nsId}/control/mci/{mciId}/vm/{vmId}:
    get:
      consumes:
      - application/json
      description: Control the lifecycle of VM (suspend, resume, reboot, terminate)
      operationId: GetControlMciVm
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - default: mci01
        description: MCI ID
        in: path
        name: mciId
        required: true
        type: string
      - default: g1-1
        description: VM ID
        in: path
        name: vmId
        required: true
        type: string
      - description: Action to MCI
        enum:
        - suspend
        - resume
        - reboot
        - terminate
        in: query
        name: action
        required: true
        type: string
      - description: Force control to skip checking controllable status
        enum:
        - "false"
        - "true"
        in: query
        name: force
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Control the lifecycle of VM (suspend, resume, reboot, terminate)
      tags:
      - '[MC-Infra] MCI Provisioning and Management'
...
  /ns/{nsId}/mci/{mciId}/vm/{vmId}/snapshot:
    post:
      consumes:
      - application/json
      description: Snapshot VM and create a Custom Image Object using the Snapshot
      operationId: PostMciVmSnapshot
      parameters:
      - description: Request body to create VM snapshot
        in: body
        name: vmSnapshotReq
        required: true
        schema:
          $ref: '#/definitions/model.TbVmSnapshotReq'
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - default: mci01
        description: MCI ID
        in: path
        name: mciId
        required: true
        type: string
      - default: g1-1
        description: VM ID
        in: path
        name: vmId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TbCustomImageInfo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Snapshot VM and create a Custom Image Object using the Snapshot
      tags:
      - '[Infra Resource] Image Management'
...
securityDefinitions:
  BasicAuth:
    type: basic
  Bearer:
    description: Type "Bearer" followed by a space and JWT token ([TBD] Get token
      in http://xxx.xxx.xxx.xxx:xxx/auth)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"