# src/core/model/customimage.go
/*
Copyright 2019 The Cloud-Barista Authors.
Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at
    http://www.apache.org/licenses/LICENSE-2.0
Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.
*/

// Package model is to handle object of CB-Tumblebug
package model

import (
	"time"
)

type CustomImageStatus string

const (
	MyImageAvailable   CustomImageStatus = "Available"
	MyImageUnavailable CustomImageStatus = "Unavailable"
)

// TbVmSnapshotReq is a struct to handle 'Create VM snapshot' request toward CB-Tumblebug.
type TbVmSnapshotReq struct {
	Name string `json:"name" example:"aws-ap-southeast-1-snapshot"`
}

type SpiderMyImageReq struct {
	ConnectionName string
	ReqInfo        struct {
		Name     string
		SourceVM string
	}
}

type SpiderMyImageInfo struct {
	IId IID // {NameId, SystemId}

	SourceVM IID

	Status CustomImageStatus // Available | Deleting

	CreatedTime  time.Time
	KeyValueList []KeyValue
}

type SpiderMyImageRegisterReq struct {
	ConnectionName string
	ReqInfo        struct {
		Name  string
		CSPId string // or, CSPid ?
	}
}

// TbCustomImageReq is a struct to handle 'Create custom image (VM snapshot)' request toward CB-Tumblebug.
type TbCustomImageReq struct {
	ConnectionName string `json:"connectionName"`
	Name           string `json:"name" validate:"required"`
	SourceVmId     string `json:"sourceVmId"`
	Description    string `json:"description"`

	// This field is for 'Register existing custom image'
	CspCustomImageId string `json:"cspCustomImageId"`
}

type TbCustomImageInfo struct {
	Namespace string `json:"namespace,omitempty" example:"default"` // required to save in RDB
	Id        string `json:"id" example:"aws-ap-southeast-1-snapshot"`
	Name      string `json:"name" example:"aws-ap-southeast-1-snapshot"`
	// uuid is universally unique identifier for the resource
	Uuid                 string            `json:"uuid,omitempty"`
	ConnectionName       string            `json:"connectionName" example:"aws-ap-southeast-1"`
	SourceVmId           string            `json:"sourceVmId" example:"aws-ap-southeast-1-1"`
	CspCustomImageId     string            `json:"cspCustomImageId,omitempty" example:"ami-06eb41e14121c550a"`
	CspCustomImageName   string            `json:"cspCustomImageName,omitempty" example:"aws-ap-southeast-1-snapshot"`
	Description          string            `json:"description"`
	CreationDate         time.Time         `json:"creationDate,omitempty" example:"2022-10-18T08:12:48Z"`
	GuestOS              string            `json:"guestOS,omitempty"` // Windows7, Ubuntu etc.
	Status               CustomImageStatus `json:"status" example:"Available"`
	KeyValueList         []KeyValue        `json:"keyValueList,omitempty"`
	AssociatedObjectList []string          `json:"associatedObjectList,omitempty" example:["/ns/default/mci/mci01/vm/aws-ap-southeast-1-1"]`
	IsAutoGenerated      bool              `json:"isAutoGenerated,omitempty"`

	// SystemLabel is for describing the Resource in a keyword (any string can be used) for special System purpose
	SystemLabel string `json:"systemLabel,omitempty" example:"Managed by CB-Tumblebug" default:""`
}